Copy 24Q3 sqlite 3.44.5 into 3.44.3 directories
Copy the 3.44.5 source into the 3.44.3 directory tree.  The sqlite
version can be verified by examining the content of the source files.

This type of change is not required after 25Q4.

Manually tested by loading the image and querying the sqlite version.

Manually tested by comparing SQLiteDatabasePerfTest metrics with the
base sqlite (3.44.3) to the new sqlite (3.44.5).

Flag: EXEMPT cve patch
Bug: 430889718
Test: atest
 * CtsDatabaseTestCases
 * FrameworksCoreTests:android.database
Change-Id: I2c640771287299c4373723da03f1cd916d3e7863